pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
version = '0.0.4'
name = "py-teamup-api"
authors = [{ name = "Tony", email = "cordial.egg1739@fastmail.com" }]
description = "A package for managing Teamup's resources."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT License" }
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
    "requests",
    "python-dotenv",
    'importlib-metadata; python_version<"3.10"',
]

[project.optional-dependencies]
test = ["pytest", "pytest-cov"]
static-code-qa = ['pre-commit', 'pytest']
release = ["build", "twine"]
dev = ["py-teamup-api[test, release, static-code-qa]"]


[tool.pytest.ini_options]
markers = ["slow: marks test as slow (deselect with '-m \"not slow:"]
